restructured to fit redux style guide
